criando pedido e inserindo produto dentro do pedido

+++ src/modules/deliveries/useCases/deleteDelivery/DeleteDeliveryUseCase.ts
@@ -1,8 +1,20 @@
 import { prisma } from "../../../../database/prismaClient";
+import { validationStatusDelivery } from "../../../../share/validators";
+import { FindByIdDeliveryUseCase } from "../findByIdDelivery/FindByIdDeliveryUseCase";
 
 
 export class DeleteDeliveryUseCase {
-    async execute(id_delivery: string, id_client: string){
-        return await prisma.deliveries.delete({ where: { id: id_delivery}})
+    async execute(id_delivery: string) {
+        const getDelivery = new FindByIdDeliveryUseCase();
+        const delivery = await getDelivery.execute(id_delivery);
+        
+        if (delivery.status === "TRANSITO") {
+            throw new Error('Pedido nao pode mais ser cancelado, pois já está em transito!');
+        }
+        if (delivery.status === "ENTREGUE") {
+            throw new Error('Pedido nao pode mais ser cancelado, pois já foi entregue!');
+        }
+
+        return await prisma.deliveries.delete({ where: { id: id_delivery } })
     }
 }
